test if JSON file exists before rmSync

+++ src/site/stages/build/drupal/static-data-files/generate.js
@@ -107,9 +107,11 @@ const writeProcessedDataFilesToCache = (
     },
   );
   for (const processedData of newSuccessful) {
-    fs.rmSync(`${fullCacheFilepath}/${processedData.filename}`, {
-      force: true,
-    });
+    if (fs.existsSync(`${fullCacheFilepath}/${processedData.filename}`)) {
+      fs.rmSync(`${fullCacheFilepath}/${processedData.filename}`, {
+        force: true,
+      });
+    }
   }
   newSuccessful.forEach(({ filename, data }) => {
     writeProcessedDataFileToCache(fullCacheFilepath, filename, data);
